chatbot/requirement.txt
google-generativeai
python-dotenv
pymssql

